Fix #70 - Permissions reset on setup-app
This commit changes the behavior of `setup-app`. Permissions are now
only set on initial call of `setup-app` and untouched on later calls,
e.g. when performing migration.

To determine whether we're currently performing an initial installation,
the current database state is used. This isn't pretty, but it should do
the job.