Disallow underscore, allow dash in instance name (fixes #143)
IE seems to have problems with setting cookies to domains with an
underscore (despite being legal) and are now disallowed - see #143.

However (sub-)domains with a dash aren't any problem and are now
allowed. We already use them in offenekommune.de (this was possible as
instances have been created through a script bypassing validation).