Update instance information in solr on a daily base (fixes #103)
In order to order instances by activity, activity information has to be
updated regularly. As of how the activity metric works, this has to be
done for all instances at once, and not only if an object in an instance
changes.

For now, this is set up on a daily base.