Fix successor relationship in Text model
Due to a mistake in the ORM mapping, the successors (children) of Text
objects have been stored, instead of the predecessors (parents), as
expected.

As it doesn't really matter whether the successor or the predecessor is
stored, as history is always linear in the current model, the
corresponding column in the database table is simply renamed from
parent_id to child_id.

Additionally, the mapper setup has been modified in order to allow for a
backref, which is now parent.

Thus, as of now, both parent and child can be accessed through the ORM.